Update views/auth/edit_user.php
The use of current_url() as form_open() argument doesn't work 
if your application is installed in a subdirectory of your 
webserver (http://mysite.com/myapp) because form_open() expects a 
uri string and, in this case, "myapp" is the folder name. I just tried it
and it produces an incorrect url and form submission doesn't work. Using
uri_string() it works fine. You should to substitute it in each form.

Thank you very much for your work!

Regards, Salvatore