Fix default group addition logic
When the ion_auth_model register() function is called with a non-empty groups array, if the default group is not included in the groups array, the new user will not be added to the default group. This runs contrary to the comment on line 862, which indicates that the new user will be added to the default group if they haven't already been added to it.